pyproject.toml
[project]
name = "webgraphai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.58.0",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.2",
]
